Refactor LocaleAwareAppCompatActivity to BaseActivity to support night mode adjust screen brightness feature.
