+++ app/src/main/java/org/mozilla/rocket/privately/PrivateModeActivity.kt
@@ -16,7 +16,7 @@ import androidx.navigation.Navigation
 import androidx.navigation.fragment.NavHostFragment
 import org.mozilla.focus.R
 import org.mozilla.focus.activity.MainActivity
-import org.mozilla.focus.locale.LocaleAwareAppCompatActivity
+import org.mozilla.focus.activity.BaseActivity
 import org.mozilla.focus.navigation.ScreenNavigator.URL_INPUT_FRAGMENT_TAG
 import org.mozilla.focus.telemetry.TelemetryWrapper
 import org.mozilla.focus.urlinput.UrlInputFragment
@@ -28,7 +28,7 @@ import org.mozilla.rocket.tabs.SessionManager
 import org.mozilla.rocket.tabs.TabViewProvider
 import org.mozilla.rocket.tabs.TabsSessionProvider
 
-class PrivateModeActivity : LocaleAwareAppCompatActivity(),
+class PrivateModeActivity : BaseActivity(),
         FragmentListener,
         TabsSessionProvider.SessionHost {
